Refactor chats.ts into separate files for chat guards, queries, mutations, and schema for chat management. Also add full-text and vector search indices
